delete duplicate files
