+++ dubbo-remoting/dubbo-remoting-api/src/main/java/org/apache/dubbo/remoting/exchange/PortUnificationExchanger.java
@@ -19,7 +19,12 @@ package org.apache.dubbo.remoting.exchange;
 import org.apache.dubbo.common.URL;
 import org.apache.dubbo.common.logger.Logger;
 import org.apache.dubbo.common.logger.LoggerFactory;
+import org.apache.dubbo.remoting.ChannelHandler;
+import org.apache.dubbo.remoting.Client;
+import org.apache.dubbo.remoting.RemotingException;
+import org.apache.dubbo.remoting.RemotingServer;
 import org.apache.dubbo.remoting.api.PortUnificationServer;
+import org.apache.dubbo.remoting.api.newportunification.PortUnificationTransporter;
 
 import java.util.ArrayList;
 import java.util.concurrent.ConcurrentHashMap;
@@ -27,31 +32,15 @@ import java.util.concurrent.ConcurrentMap;
 
 public class PortUnificationExchanger {
 
-    private static final Logger log = LoggerFactory.getLogger(PortUnificationExchanger.class);
-    private static final ConcurrentMap<String, PortUnificationServer> servers = new ConcurrentHashMap<>();
-
-    public static void bind(URL url) {
-        servers.computeIfAbsent(url.getAddress(), addr -> {
-            final PortUnificationServer server = new PortUnificationServer(url);
-            server.bind();
-            return server;
-        });
+    public RemotingServer bind(URL url, ChannelHandler handler) throws RemotingException {
+        return getTransporter(url).bind(url, handler);
     }
 
-    public static void close() {
-        final ArrayList<PortUnificationServer> toClose = new ArrayList<>(servers.values());
-        servers.clear();
-        for (PortUnificationServer server : toClose) {
-            try {
-                server.close();
-            } catch (Throwable throwable) {
-                log.error("Close all port unification server failed", throwable);
-            }
-        }
+    public Client connect(URL url, ChannelHandler handler) throws RemotingException {
+        return getTransporter(url).connect(url, handler);
     }
 
-    // for test
-    public static ConcurrentMap<String, PortUnificationServer> getServers() {
-        return servers;
+    public static PortUnificationTransporter getTransporter(URL url) {
+        return url.getOrDefaultFrameworkModel().getExtensionLoader(PortUnificationTransporter.class).getAdaptiveExtension();
     }
 }
