+++ dubbo-remoting/dubbo-remoting-netty4/src/main/java/org/apache/dubbo/remoting/transport/netty4/portunification/NettyPortUnificationTransporter.java
@@ -36,7 +36,7 @@ public class NettyPortUnificationTransporter implements PortUnificationTransport
                 if ( server == null ) {
                     server = new NettyPortUnificationServer(url, handler) ;
                 } else {
-                    server.AddNewUrl(url, handler);
+                    //todo reset or other operation
                 }
             }
         }
